fix(datepicker): fix scope leak when opening picker multiple times
Prevents creation of scopes each time datepicker is shown by using a new scope for the tipElement in tooltip service, which in the basis for the datepicker. This new scope is destroyed when the tip is hidden, thereby cleaning up resources.
Had to change other tooltip based directives custom event hook-up in show() to use $timeout instead of setTimeout to prevent errors, by delaying code execution inside the $digest cycle.