fix(typeahead): remove the ‘change’ event handlers from the typeahead element
Stop ngModel from handling ‘change’ event fired by firefox on
input loss of focus to prevent an undesired change from ngModel to $viewValue.
Previously, on firefox, if $viewValue was an object set by typeahead,
this event would change $viewValue to the string value on the input element.

Could break other directives that rely on the ’change’ event if used on the same element as typeahead (fixes #1319, #652)